refactor: extract models.py, ui.py, docx_utils.py from oversized index.py
- Extract shared model utilities (_get_cached_model_path, resolve_flashrank_model_name, configure_offline_mode) into models.py, eliminating duplication between index.py and search.py
- Extract IndexingUI, FileProcessingContext, FileProcessingTimeoutError, GracefulAbort into ui.py (~560 lines)
- Extract DOCX processing (_parse_heading_level, _get_doc_temp_dir, _convert_doc_to_docx, split_docx_into_heading_documents) into docx_utils.py (~250 lines)
- Modernize type annotations in index.py to Python 3.11+ syntax (list[], dict[], X | None)
- Add ruff linter configuration to pyproject.toml

index.py drops from 2,655 to 1,720 lines (-35%).
